Refactoring getTailwindClassName into own file
