+++ src/css/utils.ts
@@ -22,3 +22,6 @@ export const responsiveStyle = (rules: {
 
 export const selectorize = (classNames: string) =>
   '.' + classNames.split(' ').join('.')
+
+export const isTruthy = <T>(value: T | undefined | null): value is T =>
+  Boolean(value)
